Create bootstrap-ubuntu script
[Problem]
I'd like to have some documentation,
and ideally a script, that runs through
the setup of Ubuntu for me.

[Solution]
While I don't have the script working idempotently
yet, it's in a state where I can rely on it as
a reference for sequential instructions,
especially given how infrequently I'll use it.